refactor(ipc): split ssh.ts into focused modules
