+++ config/max-lines-baseline.txt
@@ -27,7 +27,6 @@ inline src/main/index.ts
 inline src/main/ipc/filesystem-watcher.ts
 inline src/main/ipc/filesystem.ts
 inline src/main/ipc/repos.ts
-inline src/main/ipc/ssh.ts
 inline src/main/ipc/worktree-remote.ts
 inline src/main/keybindings/keybinding-file.ts
 inline src/main/linear/issues.ts
